futher work on assign01

+++ assignments/assign01/src/player.java
@@ -17,7 +17,8 @@ public class player {
     }
 
     double calcRatio() {
-        return matchWins / matchLoss;
+        //typecasting necessary
+        return (double) matchWins / matchLoss;
     }
 
     void printRatio() {
